+++ src/components/Header.tsx
@@ -1,9 +1,7 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import { useRecoilValue } from 'recoil';
 import AppTheme from '../models/Theme';
-import navigationState from '../store/atoms/navigation';
 import themeState from '../store/atoms/theme';
-import { getIsLoopLine } from '../utils/loopLine';
 import CommonHeaderProps from './CommonHeaderProps';
 import HeaderJRWest from './HeaderJRWest';
 import HeaderSaikyo from './HeaderSaikyo';
@@ -20,24 +18,9 @@ const Header = ({
   lineDirection,
   stations,
   connectedNextLines,
+  isLast,
 }: CommonHeaderProps): React.ReactElement => {
   const { theme } = useRecoilValue(themeState);
-  const { trainType } = useRecoilValue(navigationState);
-
-  const isLast = useMemo(() => {
-    if (getIsLoopLine(line, trainType)) {
-      return false;
-    }
-
-    return lineDirection === 'INBOUND'
-      ? stations.findIndex((s) => s.id === nextStation?.groupId) ===
-          stations.length - 1
-      : stations
-          .slice()
-          .reverse()
-          .findIndex((s) => s.groupId === nextStation?.groupId) ===
-          stations.length - 1;
-  }, [line, lineDirection, nextStation?.groupId, stations, trainType]);
 
   switch (theme) {
     case AppTheme.TokyoMetro:
